Make long-range LERW termination and steps overflow-safe
Terminate every one- through five-dimensional long-range loop-erased walk as soon as it reaches the configured L boundary, independent of whether L appears in the legacy observable-threshold table. Record only thresholds actually crossed within the configured domain so nonstandard boundaries no longer run forever or index visited storage out of range.

Compute heavy-tailed step lengths and proposed coordinates in long double and int64_t, validate finite conversion bounds before casting, and treat any representationally huge proposal as a boundary exit. Keep stored in-domain coordinates as integers and avoid signed overflow in coordinate addition.

Erase a loop's visited tail before pop_back invalidates the element, preserving the existing loop-erasure result while removing undefined behavior. Replace compiler-specific bit-index intrinsics with checked division/modulo addressing and reject out-of-range accesses explicitly.

Harden the duplicated runtime framework across all five source trees: preserve the RNG stream using unsigned arithmetic, validate dimension/model/sampling inputs, use RAII storage, correct histogram and observable edge cases, and make output cleanup explicit.

Add a dimension-selectable build, deterministic RNG/BitArray regression tests, and a five-dimension CI matrix. Verified complete builds for 1d through 5d, sanitizer tests, and smoke runs at non-tabulated L=10. The pre-existing common.h portability conversion remains outside this commit.

+++ 2d/src/config/Configuration/Configuration.hpp
@@ -23,8 +23,8 @@ public:
     int L_max;
     int Dim;
     double Sigma;
-    std::array<double, 2> string_head; // Range for the random walk
-    std::vector<std::array<double, 2>> string;
+    std::array<int, 2> string_head; // Lattice coordinate of the walk head
+    std::vector<std::array<int, 2>> string;
     int length_string;
     int complete_length_string; // Total length of the random walk
